requirements.txt
comtypes
PyYAML
typing